// frontend/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
      action_plans: {
        Row: {
          created_at: string
          created_by: string
          date: string
          description: string | null
          id: string
          is_executed: boolean | null
          location: string
          marketing_manager_status: Database["public"]["Enums"]["action_status"]
          sales_director_status: Database["public"]["Enums"]["action_status"]
          supervisor_status: Database["public"]["Enums"]["action_status"]
          targeted_bricks: string[] | null
          targeted_delegates: string[] | null
          targeted_doctors: string[] | null
          targeted_products: string[] | null
          targeted_sales_directors: string[] | null
          targeted_supervisors: string[] | null
          type: Database["public"]["Enums"]["action_types"]
          updated_at: string
        }
        Insert: {
          created_at?: string
          created_by: string
          date: string
          description?: string | null
          id?: string
          is_executed?: boolean | null
          location: string
          marketing_manager_status?: Database["public"]["Enums"]["action_status"]
          sales_director_status?: Database["public"]["Enums"]["action_status"]
          supervisor_status?: Database["public"]["Enums"]["action_status"]
          targeted_bricks?: string[] | null
          targeted_delegates?: string[] | null
          targeted_doctors?: string[] | null
          targeted_products?: string[] | null
          targeted_sales_directors?: string[] | null
          targeted_supervisors?: string[] | null
          type?: Database["public"]["Enums"]["action_types"]
          updated_at?: string
        }
        Update: {
          created_at?: string
          created_by?: string
          date?: string
          description?: string | null
          id?: string
          is_executed?: boolean | null
          location?: string
          marketing_manager_status?: Database["public"]["Enums"]["action_status"]
          sales_director_status?: Database["public"]["Enums"]["action_status"]
          supervisor_status?: Database["public"]["Enums"]["action_status"]
          targeted_bricks?: string[] | null
          targeted_delegates?: string[] | null
          targeted_doctors?: string[] | null
          targeted_products?: string[] | null
          targeted_sales_directors?: string[] | null
          targeted_supervisors?: string[] | null
          type?: Database["public"]["Enums"]["action_types"]
          updated_at?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never